Updated Cassandraql example to use Eclipse JKube
(Fabric8-maven-plugin will soon be deprecated)
