+++ examples/camel-example-cassandra-kubernetes/src/main/resources/jkube/cassandra-service.yaml
