+++ examples/camel-example-cassandra-kubernetes/src/main/resources/jkube/cassandra-statefulset.yaml
@@ -15,13 +15,16 @@
 # limitations under the License.
 #
 
-apiVersion: "apps/v1beta1"
+apiVersion: "apps/v1"
 kind: StatefulSet
 metadata:
   name: cassandra
 spec:
   serviceName: cassandra
   replicas: 2
+  selector:
+    matchLabels:
+      app: cassandra
   template:
     metadata:
       labels:
